triangle v1

+++ core/include/Material.hpp
@@ -7,7 +7,7 @@ struct Material {
 	Vector3f color;
 	float k_diffuse;
 
-	float specular_power; // sharpness of glare
+	size_t specular_power; // sharpness of glare
 	float k_specular;
 
 	bool reflector;
@@ -22,7 +22,7 @@ struct Material {
 	Material(
 		const Vector3f& color,
 		float k_diffuse,
-		float specular_power,
+		size_t specular_power,
 		float k_specular,
 		bool reflector,
 		float k_reflective,
