+++ core/src/Material.cpp
@@ -5,7 +5,7 @@
 Material::Material(
 	const Vector3f& color,
 	float k_diffuse,
-	float specular_power,
+	size_t specular_power,
 	float k_specular,
 	bool reflector,
 	float k_reflective,
